chore(batch-scheduler): remove old scheduler impl and rename the file
