+++ src/frontend/src/scheduler/execution/stage.rs
@@ -22,12 +22,13 @@ use crate::optimizer::plan_node::PlanNodeType;
 use crate::scheduler::execution::stage::StageState::Pending;
 use crate::scheduler::execution::QueryMessage;
 use crate::scheduler::plan_fragmenter::{ExecutionPlanNode, QueryStageRef, StageId};
-use crate::scheduler::schedule::{TaskId, WorkerNodeManagerRef};
+use crate::scheduler::worker_node_manager::WorkerNodeManagerRef;
 
 // Root stage always has only one task.
 pub const ROOT_TASK_ID: u32 = 0;
 // Root task has only one output.
 pub const ROOT_TASK_OUTPUT_ID: u32 = 0;
+pub(crate) type TaskId = u32;
 
 enum StageState {
     Pending,
