+++ src/frontend/src/scheduler/mod.rs
@@ -22,8 +22,7 @@ mod execution;
 pub mod plan_fragmenter;
 mod query_manager;
 pub use query_manager::*;
-#[allow(dead_code)]
-pub mod schedule;
+pub mod worker_node_manager;
 
 /// Context for mpp query execution.
 pub struct ExecutionContext {
